app/requirements.txt
fastapi 
uvicorn 
jinja2
python-dotenv
python-multipart
reportlab
asyncpg
databases[postgresql]
WeasyPrint
